decouple video file operation logic
